+++ src/com/dakoda/alr/item/equip/weapon/Weapon.java
@@ -1,6 +1,7 @@
-package com.dakoda.alr.item.equip;
+package com.dakoda.alr.item.equip.weapon;
 import com.dakoda.alr.game.profession.Profession;
 import com.dakoda.alr.item.Item;
+import com.dakoda.alr.item.equip.EquipMaterial;
 
 import java.util.ArrayList;
