+++ src/com/dakoda/alr/item/misc/treasure/ItemTreasure.java
@@ -1,4 +1,4 @@
-package com.dakoda.alr.item.misc;
+package com.dakoda.alr.item.misc.treasure;
 import com.dakoda.alr.item.Item;
 
 import static com.dakoda.alr.item.ItemType.TREASURE;
